third_party/WebKit/LayoutTests/fast/events/mouse-events-within-no-element.html

<!DOCTYPE html>
<iframe id="sourceFrame" style="width: 400px; height: 300px;" src="resources/mouse-events-within-no-element-iframe.html"></iframe>
<p id="description"></p>
<div id="console"></div>
<script src="../../resources/js-test.js"></script>
<script>
description("Tests that the mouse event is fired when the mouse pointer stays in document but there is no element");
if (window.testRunner)
    testRunner.dumpChildFrames();
</script>
