third_party/blink/web_tests/fast/events/mouse-events-within-no-element-expected.txt

Tests that the mouse event is fired when the mouse pointer stays in document but there is no element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
POINTER

PASS events[0] is "mousemove"
PASS targets[0] is "#document"
PASS events[1] is "mousemove"
PASS targets[1] is "BODY"
PASS events[2] is "mousemove"
PASS targets[2] is "#document"
PASS events[3] is "mouseover"
PASS targets[3] is "#document"
PASS events[4] is "mouseover"
PASS targets[4] is "BODY"
PASS events[5] is "mouseover"
PASS targets[5] is "#document"
PASS events[6] is "mouseout"
PASS targets[6] is "#document"
PASS events[7] is "mouseout"
PASS targets[7] is "BODY"
PASS events[8] is "mouseout"
PASS targets[8] is "#document"
PASS events[9] is "mouseenter"
PASS targets[9] is "#document"
PASS events[10] is "mouseleave"
PASS targets[10] is "#document"
PASS events[11] is "mouseenter"
PASS targets[11] is "#document"
PASS events[12] is "mouseenter"
PASS targets[12] is "BODY"
PASS events[13] is "mouseleave"
PASS targets[13] is "BODY"
PASS events[14] is "mouseleave"
PASS targets[14] is "#document"
PASS successfullyParsed is true

TEST COMPLETE
